% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Plotly.R
\name{add_colors_names}
\alias{add_colors_names}
\title{Set colors of labels}
\usage{
add_colors_names(x, colors, pos = 1)
}
\arguments{
\item{x}{vector of labels.}

\item{colors}{named list of colors.}

\item{pos}{position of the substring to be considered.}
}
\value{
a modified vector x with colors set using HTML syntax.
}
\description{
Using a subpart of each label (split by "-" separator and take the first part), set the color using a
named list of colors.
}
\keyword{internal}
