% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Plotly.R
\name{add_symbols_pvals}
\alias{add_symbols_pvals}
\title{Add exponent star symbols on row names to show pvalues of some tests.}
\usage{
add_symbols_pvals(df_p, color, alpha_left, y_ann, y)
}
\arguments{
\item{df_p}{a data.frame with first column containing original row names and second columns containing a pvalue.}

\item{color}{hexadecimal or string color.}

\item{alpha_left}{threshold for showing 1 star.}

\item{y_ann}{vector of current row names.}

\item{y}{vector of original row names.}
}
\value{
a modified vector y_ann with additional symbols if any pvalue is below the threshold.
}
\description{
Add exponent star symbols on row names to show pvalues of some tests.
}
\keyword{internal}
