% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Plotly.R
\name{add_z_annotations_htmp}
\alias{add_z_annotations_htmp}
\title{Add text annotations on heatmap cells}
\usage{
add_z_annotations_htmp(
  z,
  fig,
  black_white_cutoff = 0.16,
  font = list(size = 8)
)
}
\arguments{
\item{z}{a numeric matrix or vector.}

\item{fig}{plotly figure containing the heatmap.}

\item{black_white_cutoff}{(optional) Cutoff for deciding on text font color.}

\item{font}{(optional) Text font parameters.}
}
\value{
the figure object \code{fig}.
}
\description{
Add text annotations on heatmap cells
}
\keyword{internal}
