% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Plotly.R
\name{complete_col_df}
\alias{complete_col_df}
\title{Complete a column of a dataframe.}
\usage{
complete_col_df(df, col, vals)
}
\arguments{
\item{df}{The dataframe.}

\item{col}{The name of the column.}

\item{vals}{Vals that should exist in the column. If values are missing, they are added as extra rows.}
}
\value{
a data.frame
}
\description{
Complete a column of a dataframe.
}
\keyword{internal}
