% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Plotly.R
\name{compute_coordinates_burden_plot}
\alias{compute_coordinates_burden_plot}
\title{Compute the coordinates of each point in the burden plot.}
\usage{
compute_coordinates_burden_plot(
  df,
  groups_order,
  col_burden,
  col_groups,
  offset = 0.1
)
}
\arguments{
\item{df}{A \code{data.frame} containing the columns \code{col_burden} and \code{col_groups}.}

\item{groups_order}{A character vector containing ordered group names.}

\item{col_burden}{Name of the column holding burden values.}

\item{col_groups}{Name of the column holding groups values.}

\item{offset}{(optional) The j^th scatter plot is drawn between j+offset and j+1-offset.}
}
\description{
Compute the coordinates of each point in the burden plot.
}
\seealso{
\code{\link[=compute_groups_order_burden_plot]{compute_groups_order_burden_plot()}}
}
