% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Plotly.R
\name{compute_groups_order_burden_plot}
\alias{compute_groups_order_burden_plot}
\title{Compute the groups order for the burden plot.}
\usage{
compute_groups_order_burden_plot(
  df,
  col_burden,
  col_groups,
  groups_keep = NULL
)
}
\arguments{
\item{df}{A \code{data.frame} containing the columns \code{col_burden} and \code{col_groups}.}

\item{col_burden}{Name of the column holding burden values.}

\item{col_groups}{Name of the column holding groups values.}

\item{groups}{(optional) Character vector containing names of groups to be retained.}
}
\value{
a character vector containing ordered group names.
}
\description{
Compute the groups order for the burden plot.
}
