% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Plotly.R
\name{configure_axes_htmp}
\alias{configure_axes_htmp}
\title{Configure x and y axes of the heatmap}
\usage{
configure_axes_htmp(
  z,
  fig,
  y_tick_font = list(size = 7),
  x_tick_font = list(size = 12),
  y_side = "left",
  x_side = "bottom",
  linewidth = 0.01
)
}
\arguments{
\item{z}{a numeric matrix or vector.}

\item{fig}{plotly figure containing the heatmap.}

\item{y_tick_font}{(optional) List with font parameters of y-axis ticks.}

\item{x_tick_font}{(optional) List with font parameters of x-axis ticks.}

\item{y_side}{(optional) Size of labels of y-axis.}

\item{x_side}{(optional) Size of labels of x-axis.}

\item{linewidth}{(optional) Linewidth of external borders (axes).}

\item{black_white_cutoff}{(optional) Cutoff for deciding on text font color.}
}
\value{
the figure object \code{fig}
}
\description{
Configure x and y axes of the heatmap
}
\keyword{internal}
