% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Plotly.R
\name{configure_yaxes_row_bar}
\alias{configure_yaxes_row_bar}
\title{Configure y-axis for bar plots}
\usage{
configure_yaxes_row_bar(
  fig,
  title = "",
  tickfont = list(size = 8),
  titlefont = list(size = 12),
  ...
)
}
\arguments{
\item{fig}{plotly figure containing the heatmap.}

\item{title}{(optional) Title of axis.}

\item{tickfont}{(optional) A list to specify tick font parameters. For instance use \code{list(size=12)} to specify
a tick font of size 12.}

\item{titlefont}{(optional) A list to specify title font parameters. For instance use \code{list(size=12)} to specify
a title font of size 12.}

\item{...}{Extra parameters added to the xaxis configuration list.}
}
\value{
the figure object \code{fig}.
}
\description{
Configure y-axis for bar plots
}
\keyword{internal}
