% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Plotly.R
\name{cumulative_scatter_plots}
\alias{cumulative_scatter_plots}
\title{Draw multiple scatter plots showing cumulative distributions}
\usage{
cumulative_scatter_plots(
  dfs,
  groups,
  stacks,
  offset,
  ytitle,
  ytitlefont = list(size = 14),
  yname = "Burden",
  yrange = c(10^-3, 10^4),
  ylabelsfont = list(size = 12),
  min_median_size = 5,
  colors_bkgd = c("#f3f3f3", "#fffff"),
  markers_scatter = list(list(size = 5, color = "black")),
  lines_median = list(list(color = "red")),
  lwd_axes = 2,
  col_id = "Subject_Id"
)
}
\arguments{
\item{dfs}{A 2-level named list of dataframes. 1st level of names are stacks, 2nd level of names are groups.}

\item{groups}{A character vector.}

\item{stacks}{A character vector.}

\item{offset}{For group i, the scatter plot is drawn between i+offset and i+1-offset.}

\item{ytitle}{Title y-axis.}

\item{ytitlefont}{Font title y-axis.}

\item{yname}{Name of the y variable. Used for the hovering template.}

\item{yrange}{Range y-axis.}

\item{ylabelsfont}{Font of y-axis labels.}

\item{min_median_size}{Minimum number of points required in the points cloud to draw the median.}

\item{colors_bkgd}{Vector of colors for the background. If less than the number of grounds, values are recycled using
\code{rep} function.}

\item{markers_scatter}{List of markers characteristics.}

\item{lines_median}{List of median line characteristics.}

\item{lwd_axes}{Linewidth of line delimiting y anx x axes.}

\item{col_id}{Name of column id for hovering.}
}
\description{
Scatter plots are organized by groups and stacks. The groups define the numbers of subplots while stacks define
the number of points clouds in each subplot.
}
\author{
Yoann Pradat
}
