% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Plotly.R
\name{draw_barplot_above_htmp}
\alias{draw_barplot_above_htmp}
\title{Draw a stacked barplot to be integrated above a heatmap.}
\usage{
draw_barplot_above_htmp(
  df_plot,
  stacks2colors,
  fonts,
  linewidth,
  x = NULL,
  x_ann = NULL,
  xaxis_showticklabels = F,
  legend_title = "",
  legend_x = c(-0.2, -0.2),
  legend_y = c(0.1, 0.9),
  legend_x_to_y_ratio = 0.3,
  legend_y_gap = 0.06,
  legend_x_gap = NULL,
  legend_orientation = "v"
)
}
\arguments{
\item{df_plot}{a data.frame with the stacks defined in the first column, the column labels defined in the column
\code{col_x} and the values to be stacked in the \code{col_y} column.}

\item{stacks2colors}{a named list of colors.}

\item{fonts}{(optional) List of fonts. Required names are 'x_tick_heatmap', 'y_tick_heatmap', 'x_tick_rowbar'
and 'legend'. Each font is a list with font parameters, e.g, 'size'.}

\item{linewidth}{(optional) Linewidth of external borders (axes).}

\item{x}{a vector of ordered x-axis labels matching values from \code{col_x} column.}

\item{x_ann}{a vector of x-axis modified labels to be displayed as x label names.}

\item{xaxis_showticklabels}{set to T to display x label names (using \code{x_ann} if not NULL or \code{x} otherwise).}

\item{legend_title}{title for the legend..}

\item{legend_x}{domain along the x-axis of the barplot in paper ref where the legend will be drawn.}

\item{legend_y}{domain along the y-axis of the barplot in paper ref where the legend will be drawn.}

\item{legend_x_to_y_ratio}{ratio controlling the aspect of legend items.}

\item{legend_y_gap}{gap between legend items along the y-axis.}

\item{legend_x_gap}{gap between legend items along the x-axis.}

\item{legend_orientation}{choose 'h' or 'v'.}

\item{y}{a vector of ordered y-axis labels.}
}
\value{
a plotly figure.
}
\description{
Draw a stacked barplot to be integrated above a heatmap.
}
\author{
Yoann Pradat
}
\keyword{internal}
