% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Plotly.R
\name{draw_burden_plot}
\alias{draw_burden_plot}
\title{Draw multiple scatter plots showing cumulative distributions}
\usage{
draw_burden_plot(
  dfs,
  groups,
  stacks,
  offset,
  stacks2colors,
  marker_size = 4,
  marker_opacity = 1,
  median_width = 2,
  yname = "Burden",
  yrange = c(10^-3, 10^4),
  min_median_size = 5,
  ytitle = "Burden",
  margin = list(t = 250),
  lwd_axes = 2,
  ytitlefont = list(size = 18),
  xlabelsfont = list(size = 20),
  ylabelsfont = list(size = 20),
  legendfont = list(size = 18),
  col_id = "Subject_Id"
)
}
\arguments{
\item{dfs}{A 2-level named list of dataframes. 1st level of names are stacks, 2nd level of names are groups.}

\item{groups}{A character vector.}

\item{stacks}{A character vector.}

\item{offset}{For group i, the scatter plot is drawn between i+offset and i+1-offset.}

\item{stacks2colors}{A named list with names being values in \code{stacks} and values being color names.}

\item{marker_size}{Size of the scatter markers.}

\item{median_width}{Width of the median line.}

\item{yname}{Name of the y variable. Used for the hovering template.}

\item{yrange}{Range y-axis.}

\item{min_median_size}{Minimum number of points required in the points cloud to draw the median.}

\item{ytitle}{Title y-axis.}

\item{margin}{Named list passed to the \code{margin} parameter of \code{plotly::layout}.}

\item{lwd_axes}{Linewidth of line delimiting y anx x axes.}

\item{ytitlefont}{Font title y-axis.}

\item{xlabelsfont}{Font of the x-axis labels.}

\item{ylabelsfont}{Font of y-axis labels.}

\item{legendfont}{Font of the legend.}

\item{col_id}{Name of column id for hovering.}
}
\description{
Scatter plots are organized by groups and stacks. The groups define the numbers of subplots while stacks define
the number of points clouds in each subplot.
}
\author{
Yoann Pradat
}
