% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Plotly.R
\name{draw_facetted_heatmap_barplots_1}
\alias{draw_facetted_heatmap_barplots_1}
\title{Draw a facetted figure with template 1.}
\usage{
draw_facetted_heatmap_barplots_1(
  dfs,
  col_var,
  names2plots,
  names2colors,
  colors_palette_heatmap = "Reds",
  colors_limits_heatmap = NULL,
  col_pval = "p.value",
  width_one = 100,
  height_one = 10,
  alpha_left = 0.1,
  alpha_heatmap = 0.1,
  col_stack = NULL,
  width_edges_heatmap = 0.05,
  fonts = list(z_heatmap = list(size = 8), x_tick_heatmap = list(size = 10),
    y_tick_heatmap = list(size = 6), x_tick_row_bar = list(size = 8), x_title_row_bar =
    list(size = 10), cohort_size = list(size = 10), legend = list(size = 9)),
  bargap = 0.1,
  stacks2colors = NULL,
  showlegend = TRUE,
  add_colors_names = TRUE,
  x_tick_heatmap_side = "bottom",
  x_title_row_bar = "",
  add_cohorts_sizes = F
)
}
\arguments{
\item{dfs}{a named list of dataframes containing the values indicated in the list \code{names2plots}.}

\item{col_var}{A character vector defining the field used as variables.}

\item{names2plots}{a list with names among 'left_pvals', 'heatmap', 'heatmap_hover', 'heatmap_pvals',
'middle_right_bar', 'extreme_right_bar'. The corresponding values must be in the list of names of \code{dfs}.}

\item{names2colors}{a list with names that must match values of the vectors names2plots$left_pvals and
names2plots$heatmap_pvals.}

\item{colors_palette_heatmap}{(optional) The color palette to be used for the heatmap. You may use "Reds", "Blues"
or a list of colors that matches the length of color_limits_heatmap.}

\item{colors_limits_heatmap}{(optional) The limit values for deciding on cell colors in the heatmap.}

\item{width_one}{(optional) The width of the plot is obtained as \code{width_one} multiplied by the number of columns
in \code{z}.}

\item{height_one}{(optional) The height of the plot is obtained as \code{height_one} multiplied by the number of rows
in \code{z}.}

\item{alpha_left}{(optional) Threshold for drawing the bar for a significant variable pvalue.}

\item{alpha_heatmap}{(optional) Threshold for highlighting edges of a significant cell pvalue.}

\item{col_stack}{(optional) A character vector defining the field used for the extreme right stack barplot. Used
only if 'extreme_right_bar' in the names of \code{names2plots}.}

\item{width_edges_heatmap}{(optional) Width of the line defining the edges of the heatmap.}

\item{fonts}{(optional) List of fonts. Required names are 'x_tick_heatmap', 'y_tick_heatmap', 'x_tick_rowbar'
and 'legend'. Each font is a list with font parameters, e.g, 'size'.}

\item{bargap}{(optional) Spacing between bars of barplot.}

\item{stacks2colors}{(optional) A list of colors with names corresponding to values for the extreme right stack
barplot. Used only if 'extreme_right_bar' in the names of \code{names2plots}.}

\item{showlegend}{(optional) Set to FALSE to hide the legend.}

\item{add_colors_names}{(optional) Set to FALSE to not use colors in x-axis labels.}

\item{x_tick_heatmap_side}{(optional) Choose 'bottom' or 'top'.}

\item{x_title_row_bar}{(optional) Title for the x-axis of the side bar plot.}

\item{add_cohorts_sizes}{(optional) Set to TRUE to show cohort sizes above or below x tick labels.}
}
\value{
a list of plotly figure objects.
}
\description{
The main plot in the figure is the central heatmap. This heatmap serves to show the distribution of counts or
percentages of an event in a fixed number of variables and fixed number of samples.

The figure may additionally contain a
\itemize{
\item{a left double inverted barplot} {it is used to display -log10(pvalues) distribution from
a per-variable test (e.g Fisher stratified by the samples).}
\item{a middle right barplot} {it is used to display total variable counts.}
\item{an extreme right barplot} {it is used to display additional information about a variable in the form of a
stacked barplot. It is useful to show the breakdown of a total variable count into types.}
}
}
\author{
Yoann Pradat
}
