% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Plotly.R
\name{draw_facetted_heatmap_barplots_2}
\alias{draw_facetted_heatmap_barplots_2}
\title{Draw a facetted figure with template 2.}
\usage{
draw_facetted_heatmap_barplots_2(
  dfs,
  col_var,
  names2plots,
  names2colors,
  linewidth = 2,
  width_one = 100,
  height_one = 10,
  alpha_left = 0.1,
  alpha_heatmap = 0.1,
  width_edges_heatmap = 0.05,
  legend_titles = c("ESCAT levels", "# of actionable<br>alterations"),
  legend_x = -0.25,
  legend_x_to_y_ratio = 0.2,
  legend_y_gap = 0.06,
  fonts = list(z_heatmap = list(size = 8), x_tick_heatmap = list(size = 10),
    y_tick_heatmap = list(size = 10), x_tick_row_bar = list(size = 10), y_title_row_bar =
    list(size = 12), cohort_size = list(size = 10), legend = list(size = 12),
    legend_label = list(size = 12), legend_title = list(size = 16)),
  add_colors_names = TRUE
)
}
\arguments{
\item{dfs}{a named list of dataframes containing the values indicated in the list \code{names2plots}.}

\item{col_var}{A character vector defining the field used as variables.}

\item{names2plots}{a list with names among 'heatmap', 'heatmap_color', 'heatmap_hover', 'left_pvals',
'heatmap_pvals', 'barplot_top', 'barplot_bot'. The corresponding values must be in the list of names of \code{dfs}.}

\item{names2colors}{a list with names that must match values of the vectors names2plots$left_pvals and
names2plots$heatmap_pvals.}

\item{linewidth}{(optional) Linewidth of external borders (axes).}

\item{width_one}{(optional) The width of the plot is obtained as \code{width_one} multiplied by the number of columns
in \code{z}.}

\item{height_one}{(optional) The height of the plot is obtained as \code{height_one} multiplied by the number of rows
in \code{z}.}

\item{alpha_left}{(optional) Threshold for drawing the bar for a significant variable pvalue.}

\item{alpha_heatmap}{(optional) Threshold for highlighting edges of a significant cell pvalue.}

\item{width_edges_heatmap}{(optional) Width of the line defining the edges of the heatmap.}

\item{legend_titles}{(optional) Titles for legends left of barplots.}

\item{legend_x}{(optional) Position on the x-axis of the legend next to the heatmap barplots.}

\item{legend_x_to_y_ratio}{(optional) Ratio for controlling the aspect of legend items.}

\item{legend_y_gap}{(optional) For controlling vertical gap between legend items.}

\item{fonts}{(optional) List of fonts. Required names are 'x_tick_heatmap', 'y_tick_heatmap', 'x_tick_rowbar'
and 'legend'. Each font is a list with font parameters, e.g, 'size'.}

\item{add_colors_names}{(optional) Set to FALSE to not use colors in x-axis labels.}

\item{showlegend}{(optional) Set to FALSE to hide the legend.}
}
\value{
a list of plotly figure objects.
}
\description{
Draw a facetted figure having a heatmap in its center. This heatmap serves to show the distribution of counts or
percentages of events (rows) per groups of observations (columns). The figure may additionally contain barplots
positioned above or below the heatmap.
}
\author{
Yoann Pradat
}
