% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Plotly.R
\name{draw_legend_manual}
\alias{draw_legend_manual}
\title{Draw the legend manually}
\usage{
draw_legend_manual(
  labels2colors,
  x_min,
  x_max,
  y_min,
  y_max,
  x_gap = NULL,
  y_gap = NULL,
  orientation = "v",
  x_to_y_ratio = 1,
  xref = "paper",
  yref = "paper",
  font_labels = list(size = 14)
)
}
\arguments{
\item{labels2colors}{named list of legend colors and labels.}

\item{x_min}{x position at where legend should start.}

\item{x_max}{x position at where legend should end.}

\item{y_min}{y position at where legend should start.}

\item{y_max}{y position at where legend should end.}

\item{x_gap}{gap along x-axis between 2 legend items.}

\item{y_gap}{gap along y-axis between 2 legend items.}

\item{orientation}{choose "v" for vertical and "h" for horizontal.}

\item{x_to_y_ratio}{ratio for controlling the aspect of legend rectangles.}

\item{xref}{see shapes parameters in plotly.}

\item{yref}{see shapes parameters in plotly.}

\item{font_labels}{names list of font characteristics.}
}
\value{
a list of shapes and annotations representing the legend.
}
\description{
Draw the legend manually
}
\author{
Yoann Pradat
}
\keyword{internal}
