% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Plotly.R
\name{draw_numbered_heatmap}
\alias{draw_numbered_heatmap}
\title{Draw a simple heatmap with numbers}
\usage{
draw_numbered_heatmap(
  df_z,
  z_name,
  colors_limits = NULL,
  colors_palette = "Reds",
  width_one = 50,
  height_one = 15,
  black_white_cutoff = 0.5,
  font = list(size = 8)
)
}
\arguments{
\item{df_z}{a data.frame with numeric values.}

\item{z_name}{Name of the values for the hovering template.
'middle_right_bar', 'extreme_right_bar'. The corresponding values must be in the list of names of \code{dfs}.}

\item{colors_limits}{(optional) A numeric vector, used to associate to each unique value in \code{z} an interval.}

\item{colors_palette}{(optional) A name ("Reds" or "Blues") or a character vector of color names/codes.}

\item{width_one}{(optional) The width of the plot is obtained as \code{width_one} multiplied by the number of columns
in \code{z}.}

\item{height_one}{(optional) The height of the plot is obtained as \code{height_one} multiplied by the number of rows
in \code{z}.}

\item{black_white_cutoff}{(optional) Cutoff for deciding on text font color.}

\item{font}{(optional) Text font parameters.}
}
\value{
a plotly figure object.
}
\description{
Draw a simple heatmap with numbers
}
\author{
Yoann Pradat
}
