% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ComplexHeatmap.R
\name{draw_upset_plot}
\alias{draw_upset_plot}
\title{Draw Upset plot with predefined them}
\usage{
draw_upset_plot(
  m,
  row_annot_fontsize = 8,
  pt_size = grid::unit(3, "mm"),
  lwd = 2,
  height_top_annot = 3,
  width_set_size = 3,
  margin_row_text = 4,
  ...
)
}
\arguments{
\item{m}{object produced by ComplexHeatmap::make_comb_mat}

\item{row_annot_fontsize}{(optional) fontsize of rows annotations}

\item{pt_size}{(optional) size for dots representing combination sets}

\item{lwd}{(optional) line width for the combination sets}

\item{height_top_annot}{(optional) height in inches}

\item{width_set_size}{(optional) height in inches}

\item{margin_row_text}{(optional) width in inches of the margin surrounding set names}

\item{...}{Extra parameters passed to ComplexHeatmap::UpSet.}
}
\description{
Draw an upset plot using the function \code{ComplexHeatmap::Upset}. It takes as input
an object produced by \code{ComplexHeatmap::make_comb_mat} and draws the plot
with a predefined theme.
}
\author{
Yoann Pradat
}
