% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{get_label_colors}
\alias{get_label_colors}
\title{Associate colors to labels}
\usage{
get_label_colors(labels, pal = "Dark2", labels_unique = NULL, alpha = 1)
}
\arguments{
\item{labels}{vector.}

\item{pal}{(optional) name of a palette of RColorBrewer.}

\item{labels_unique}{(optional) If not NULL, used to defined colors for labels that are not in the labels vector.
Useful when you wish to harmonize colors between plots that do not have all the labels each.}

\item{alpha}{(optional) double in [0,1].}
}
\value{
a character vector of colors of same size as labels.
}
\description{
Takes as input a vector of labels (may be non unique) and returns a vector of the same size with one color for each
unique label in the labels. The set of unique labels may be specified. It is useful if not all labels are present in
the vector of labels.
}
\author{
Yoann Pradat
}
