% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Plotly.R
\name{get_tables_for_facetted_heatmap_barplots}
\alias{get_tables_for_facetted_heatmap_barplots}
\title{Prepare tables for the facetted heatmap barplots}
\usage{
get_tables_for_facetted_heatmap_barplots(
  df_evt_count,
  df_tt_count,
  tt_keep,
  col_evt,
  col_tt,
  col_evt_classes = NULL
)
}
\arguments{
\item{df_evt_count}{The event counts aggreagted by tumor type.}

\item{df_tt_count}{Table of tumor type counts. It must contain the columns \code{Count} and col_tt.}

\item{tt_keep}{List of tumor types included in the plot.}

\item{col_evt}{Name of the column containing the events.}

\item{col_tt}{Name of the column containing tumor types.}

\item{col_evt_classes}{(optional) Name of the column containing subclasses of events.}
}
\value{
a list of tables.
}
\description{
Return a list of dataframes with events counts and percentages.
}
