% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ComplexHeatmap.R
\name{make_upset_m}
\alias{make_upset_m}
\title{Build a \code{ComplexHeatmap::Heatmap-class} object from a dataframe}
\usage{
make_upset_m(df, field_set, field_identifier)
}
\arguments{
\item{df}{dataframe with one field for identifiers and one field to group identifiers}

\item{field_set}{Name of the field identifying groups.}

\item{field_identifier}{Name of the field identifier entries.}
}
\value{
m a \code{ComplexHeatmap::Heatmap-class} object
}
\description{
Apply the \code{ComplexHeatmap::make_comb_mat} on a named list prepared from
a dataframe.
}
\author{
Yoann Pradat
}
