% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{rect_plot_colors}
\alias{rect_plot_colors}
\title{Display a list of colors}
\usage{
rect_plot_colors(line = NULL, col = NULL, colors, cex = 1)
}
\arguments{
\item{line}{How many lines in the facetted plot.}

\item{col}{How many columns in the facetted plot.}

\item{colors}{Named list of colors.}

\item{cex}{Font size.}
}
\description{
From a named list of colors, draw a set of rectangles showing the color and the name inside the rectangle.
}
\author{
Yoann Pradat
}
