% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{render_table}
\alias{render_table}
\title{Render a table using DT}
\usage{
render_table(
  df,
  caption,
  full = F,
  nrows = 5,
  extensions = c("Buttons", "Responsive"),
  buttons = c("copy", "csv", "excel")
)
}
\arguments{
\item{df}{data.frame to render}

\item{caption}{the table caption}

\item{full}{(optional) if T, render the full table}

\item{nrows}{(optional) number of rows to render. Ignored if full=T.}

\item{extensions}{(optional) passed to parameter 'extensions' of DT::datatable.}

\item{buttons}{(optional) passed to parameter 'options' of DT::datatable. Supported values are "copy", "csv" and
"excel".}
}
\description{
Render a table using DT
}
\author{
Yoann Pradat
}
