% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprismtools.R
\docType{package}
\name{rprismtools}
\alias{rprismtools}
\title{rprismtools: A package with util functions.}
\description{
'rprismtools' provides small functions useful for making plots, for rendering html widgets or for performing common
tasks.
}
\author{
Yoann Pradat
}
