% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Plotly.R
\name{select_in_plot_evt}
\alias{select_in_plot_evt}
\title{Get the union of values that meet threshold requirements.}
\usage{
select_in_plot_evt(
  dfs_plot,
  min_counts_evt,
  col_evt,
  max_evt = NULL,
  max_evt_cohort = NULL
)
}
\arguments{
\item{dfs_plot}{List of dataframes.}

\item{min_counts_evt}{List of minimum counts for each cohort.}

\item{col_evt}{Name of the column from which to extract a entries.}

\item{max_evt}{An integer giving the maximum number of events to be considered.}

\item{max_evt_cohort}{A name identifying the table in \code{dfs_plot} to be used for selecting the events to be kept
in case there are more than \code{max_evt} selected after filtering by \code{min_counts_evt}.}
}
\value{
a vector.
}
\description{
Get the union of values that meet threshold requirements.
}
