% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Plotly.R
\name{select_vals_in_dfs}
\alias{select_vals_in_dfs}
\title{Subset a list of dataframes on a column.}
\usage{
select_vals_in_dfs(dfs, col, vals)
}
\arguments{
\item{dfs}{List of dataframes.}

\item{col}{The name of the column.}

\item{vals}{Vals that should exist in the column. If values are missing, they are added as extra rows.}
}
\value{
a data.frame
}
\description{
Subset a list of dataframes on a column.
}
