% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Plotly.R
\name{shapes_cell_edges}
\alias{shapes_cell_edges}
\title{Build list of shapes}
\usage{
shapes_cell_edges(n_row, n_col, color, width = 1)
}
\arguments{
\item{n_row}{Number of rows of the heatmap.}

\item{n_col}{Number of columns of the heatmap.}

\item{color}{Line color of the rectangle.}

\item{width}{Line width of the rectangle.}
}
\value{
a list of shapes
}
\description{
Draw a rectangular shape with a specific color for highlighting cell edges.
}
\keyword{internal}
