% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Plotly.R
\name{shapes_pvalues}
\alias{shapes_pvalues}
\title{Build list of shapes}
\usage{
shapes_pvalues(dfs_pval, names, names2colors, alpha = 0.05, width = 0.5)
}
\arguments{
\item{names}{Names of the list of data.frames.}

\item{names2colors}{Named list of colors.}

\item{alpha}{pvalue threshold for deciding to add a shape or not.}

\item{width}{Line width of the shape}

\item{dfs_pvals}{Names list of data.frames.}
}
\value{
a list of shapes
}
\description{
Draw a shape with a specific color according to the pvalue.
}
\keyword{internal}
